Add files via upload
Hangman is a word-guessing game in C++ where one player thinks of a word, and the other player guesses letters one by one. If a guessed letter is not in the word, a part of the hangman is drawn. The game continues until the word is guessed or the hangman drawing is completed.